Create a new RX_HANDLER_DROPPED to apply backpressure
This create a new value that can be forwarded up to
the driver. The driver can then stop processing right away
(at least) and, in this patch also prevent reenabling interrupt
at the end of the napi loop, until the application decides
it is time to reenable them (using ndo_reenable_itr).

This provides the roots for a generic way to apply backpressure, this
is only a proof of concept. It does not handle multi-queue, and does
not support multi-processing (I think). The backpressure int
in particular should be protected. With multi-thread there
is a race condition where the interrutps could be re-enabled before
being disabled.

When an application using the rx_handler is pinned on the same core,
receiving 40G of 64bytes traffic, it nearly nevers run and handle
<50MBps. With this it handles 9.9Gbps in the very same environment.

See github.com/tbarbette/click/tree/kpressure for an example using
the Click Modular Router.